requirements_main.txt
accelerate==1.9.0
datasets==3.6.0
hydra-core==1.3.2
lightning_bolts==0.7.0
matplotlib==3.10.5
numpy==2.3.2
omegaconf==2.3.0
packaging==25.0
PyYAML==6.0.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
zstandard==0.24.0
wandb